split create-functional-component
